API 가독성 정리

+++ InstaClone/src/main/java/com/skhkim/instaclone/service/PostServiceImpl.java
@@ -1,12 +1,14 @@
 package com.skhkim.instaclone.service;
 
+import com.skhkim.instaclone.context.LoginContext;
 import com.skhkim.instaclone.dto.PostDTO;
-import com.skhkim.instaclone.request.PostPageRequest;
+import com.skhkim.instaclone.dto.UserInfoDTO;
 import com.skhkim.instaclone.entity.Post;
 import com.skhkim.instaclone.entity.PostImage;
 import com.skhkim.instaclone.repository.PostImageRepository;
 import com.skhkim.instaclone.repository.PostRepository;
 import com.skhkim.instaclone.repository.ReplyRepository;
+import com.skhkim.instaclone.request.PostPageRequest;
 import com.skhkim.instaclone.response.PostResponse;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
@@ -36,7 +38,7 @@ public class PostServiceImpl implements PostService {
 
     @Override
     @Transactional
-    public void register(PostDTO postDTO){
+    public boolean insert(PostDTO postDTO){
 
         Map<String, Object> entityMap = EntityMapper.dtoToEntity(postDTO);
         Post post = (Post) entityMap.get("post");
@@ -45,54 +47,64 @@ public class PostServiceImpl implements PostService {
         postRepository.save(post);
         postImageRepository.saveAll(postImageList);
 
+        return true;
     }
 
     @Override
     @Transactional
-    public void modifyTitle(PostDTO postDTO){
-        Map<String, Object> entityMap = EntityMapper.dtoToEntity(postDTO);
-        Post post = (Post) entityMap.get("post");
-        postRepository.save(post);
+    public boolean update(PostDTO postDTO){
+        UserInfoDTO loginUserDTO = LoginContext.getUserInfo();
+        if(postRepository.checkValidation(postDTO.getPno(), loginUserDTO.getUserEmail())) {
+            Map<String, Object> entityMap = EntityMapper.dtoToEntity(postDTO);
+            Post post = (Post) entityMap.get("post");
+            postRepository.save(post);
+            return true;
+        }
+        else
+            return false;
     }
 
     @Override
-    public PostResponse getList(PostPageRequest postPageRequest, String name){
+    public PostResponse selectPostList(PostPageRequest postPageRequest, String name){
         Pageable pageable = postPageRequest.getPageable();
-        Slice<Post> result = postRepository.getListPage(pageable, name);
+        Slice<Post> result = postRepository.selectPostList(pageable, name);
         List<PostDTO> postDTOS = result.stream().map(EntityMapper::entityToDTO).toList();
 
         return new PostResponse(postDTOS, result.hasNext());
     }
     @Override
-    public Long getPostNumber(String email){
+    public Long selectPostNumber(String email){
         return postRepository.getPostCount(email);
-    }
-    @Override
-    public String getEmailByUserName(String userName){
-        return postRepository.getEmail(userName);
     }
 
     @Override
-    public PostDTO getPostWithAllImage(Long postId){
+    public PostDTO selectPostDetail(Long postId){
         Post post = postRepository.selectPost(postId);
         return EntityMapper.entityToDTO(post);
     }
     @Override
     @Transactional
-    public void removePost(Long pno){
-        replyRepository.deleteByPostPno(pno);
-        List<PostImage> postImageList = postImageRepository.findByPno(pno);
-        postImageRepository.deleteByPostPno(pno);
-        postImageList.forEach(postImage -> {
-            try {
-                Path path = Paths.get(uploadPath + postImage.getPath() + "/" + postImage.getUuid() + "_" + postImage.getImgName());
-                Path path_s = Paths.get(uploadPath + postImage.getPath() + "/s_" + postImage.getUuid() + "_" + postImage.getImgName());
-                Files.deleteIfExists(path);
-                Files.deleteIfExists(path_s);
-            }catch (IOException e) {
-                e.printStackTrace();
-            }
-        });
-        postRepository.deleteByPno(pno);
+    public boolean delete(Long pno){
+
+        UserInfoDTO loginUserDTO = LoginContext.getUserInfo();
+        if(postRepository.checkValidation(pno, loginUserDTO.getUserEmail())) {
+            replyRepository.deleteByPostPno(pno);
+            List<PostImage> postImageList = postImageRepository.findByPno(pno);
+            postImageRepository.deleteByPostPno(pno);
+            postImageList.forEach(postImage -> {
+                try {
+                    Path path = Paths.get(uploadPath + postImage.getPath() + "/" + postImage.getUuid() + "_" + postImage.getImgName());
+                    Path path_s = Paths.get(uploadPath + postImage.getPath() + "/s_" + postImage.getUuid() + "_" + postImage.getImgName());
+                    Files.deleteIfExists(path);
+                    Files.deleteIfExists(path_s);
+                } catch (IOException e) {
+                    e.printStackTrace();
+                }
+            });
+            postRepository.deleteByPno(pno);
+            return true;
+        }
+        else
+            return false;
     }
 }
